+++ examples/rest.php
@@ -20,6 +20,9 @@ require_once dirname(__DIR__).'/vendor/autoload.php';
 use fkooman\Http\JsonResponse;
 use fkooman\Rest\Service;
 use fkooman\Http\Exception\BadRequestException;
+use fkooman\Rest\ExceptionHandler;
+
+ExceptionHandler::register();
 
 $service = new Service();
 $service->get(
@@ -36,7 +39,6 @@ $service->get(
         return $response;
     }
 );
-
 $service->post(
     '/hello/:str',
     function ($str) {